[Common] Some fb info can be moved to profile component
There are some bugs that are not app-breaking but needs to work on; bug
list below. Changes mostly in fbLogin - we can now pull info from
facebook profile, and it gets sent to the database under users' facebook
id.  The info is then stored in the Profile component under router flux.

BUGS: When the app is launched, at least in the Android version it opens
the card page for a split second and goes straight to the profile page.
Haven't found out why yet.

We still need to import large pictures from FB and have them saved on
Firebase so we can use them.

TODO: Import more info than name and the profile pic, such as the user's
other pictures. Fix bugs. Have the profile page actually be updated with
all the info. Add edit bio functionality.