[Android] Implementing event creation - card flow
Note: Temporarily disabled passing fb information to the Profile
component at launch because it autodirects to profile for a split second
and re-renders the home component, which creates an error when trying to
populate the cards

Currently, the flow of Creating Event -> Creates a new entry in Events
database -> Imports each events as cards works.

TODO: When user swipes right, save the user's ID as the guest of the
event, and have user's profile react to the change, which would complete
the flow. Also, fix bug where the cards do not populate at launch, and
pass facebook info to the profile