[Android] Card matching
When the user swipes right on a card, their ID is now logged on the
event as 'guest'.
TODO: Instead of having a single guest entry for each of the events,
make it an array, and add another field of 'accepted_guest' where the
event creator can finish the matching flow by accepting one of the
guests.